Centralize a bunch of estimates related information and utils in src/models/estimates
