+++ client/src/models/businessConstants.ts
@@ -3,6 +3,8 @@ import _ from "lodash";
 import { assign_to_dev_helper_namespace } from "src/core/assign_to_dev_helper_namespace";
 import { lang } from "src/core/injected_build_constants";
 
+import { estimates_docs } from "./estimates";
+
 interface BusinessConceptTextValue {
   text: string;
 }
@@ -605,44 +607,6 @@ const NA_values = [
   "n/a",
   "Non-EX",
 ];
-
-const estimates_docs = {
-  MAINS: {
-    en: "Main Estimates",
-    fr: "Budget principal",
-    order: 0,
-  },
-  MYA: {
-    en: "Multi Year Appropriations",
-    fr: "Crédits disponibles des précédents exercices",
-    order: 1,
-  },
-  VA: {
-    en: "Voted Adjustments",
-    fr: "Réajustement votés",
-    order: 1.1,
-  },
-  SA: {
-    en: "Statutory Adjustments",
-    fr: "Réajustements législatifs",
-    order: 1.2,
-  },
-  SEA: {
-    en: "Supp. Estimates A",
-    fr: "Budget supp. A",
-    order: 2,
-  },
-  SEB: {
-    en: "Supp. Estimates B",
-    fr: "Budget supp. B",
-    order: 3,
-  },
-  SEC: {
-    en: "Supp. Estimates C",
-    fr: "Budget supp. C",
-    order: 4,
-  },
-};
 
 const businessConstants = {
   sos,
